requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-dotenv
black
ruff
pytest
python-jose[cryptography]
httpx
bcrypt